Update README to reflect dependency changes.
This updates the README to reflect the differences in installation due to the out-of-date repository and incomplete dependency list for Debian. It's not clear how the original installation guide was suggesting installing the packages from the dhuggins PPA, but downloading the .deb files over HTTPS is really the only way to do that now that apt-secure makes using out-of-date PPAs so difficult.